SMHA separation
