+++ src/constants/Constants.java
@@ -3,6 +3,6 @@ package constants;
 public class Constants {
 	
 	public static final Double w1 = 5.0;
-	public static final Double w2 = 100.0;
-	public static final int NumberOfInadmissibleHeuristicsForSMHAStar = 2;
+	public static final Double w2 = 10.0;
+	public static final int NumberOfInadmissibleHeuristicsForSMHAStar = 20;
 }
